Reorganize the bindings a bit
